Melhora na perfomace e resolvendo gambiarras

+++ api/Models/GrupoDeIps.cs
@@ -5,6 +5,6 @@ namespace Api.Models
     public class GrupoDeIps
     {
         public List<Ip> ListaDeIps { get; set; } = new List<Ip>();
-        public long Tempo { get; set; } = 0;
+        public double Tempo { get; set; } = 0;
     }
 }
